executor: Refuse to dequeue new job when len(vms) >= numHandlers
